+++ packages/orq-rc/RELEASES.md
@@ -12519,3 +12519,13 @@ Based on:
 - [typescript v4.16.0-rc.7] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.7] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.7 - packages/orq-rc
+
+## 2026-09-21 06:32:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.798.0 (2.938.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.8] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.8 - packages/orq-rc